Require a signed content when deleting an envelope
The previous implementation of delete had a problem because it allowed
anyone who knew the public part of the keypair used to create the
envelope to delete it. Now we require some signed content to be sent
alongside the public key to verify that the signature is valid.

This also impplies some changes in the parameters and the deletion
logic.