Set metadata community as a top level namespace
From now on, the metadata community resides at the top level of the URL
hierarchy, as in `/api/:metadata_community/envelopes`.

Its use is optional, so the bare URLs continue to  work, although a
workaround is needed because Grape does not support mounting the same
API to different routes.

Unfortunately adding a dynamic namespace seems to break grape-swagger,
so the automated generation is disabled until a fix is found.