Consider envelope/metadata community in endpoints
The envelope comunity acts as a top level contraint when dealing
with envelopes. When publishing a new envelope it will take into
account the chosen community or set the default one (Learning Registry
at the moment) if not specified. All other actions, like retrieving or
deleting envelopes, are also affected by the elected community.

This commit also renames 'metadata community' to 'envelope community' to
avoid confusion.

Closes #15.