Validate learning resources against a JSON schema
Every community has a matching JSON schema defining its correct
representation that is applied against the received resource to verify
its validity.

For now, the schemas for Learning Registry and Credential Registry just
perform a trivial validation, but the application logic is already in
place, and should work the same when the final schemas are available.